Fix #2509 make DatePicker use disabled prop
Previous to this change the disabled prop on DatePicker only worked on
component mounting, this is the property was set initially but when
later changed the component remained in the previous state.

This commit also removes two calls to `setState` on `componentDidMount`
which prevents an innecesary call to `render` on mounting.